+++ tests/test_parser_local.py
@@ -3,7 +3,7 @@ Testes do ParserLocal
 """
 from unittest import mock
 from random import randrange, randint
-from vmm_manager.app.inventario import ParserLocal
+from vmm_manager.parser.parser_local import ParserLocal
 from tests.base import Base
 from tests.dados_teste import DadosTeste
 
@@ -21,7 +21,7 @@ class TestParserLocal(Base):
         assert status is False
         assert msg == 'Arquivo de inventário vazio.'
 
-    @mock.patch('vmm_manager.app.inventario.ParserLocal._ParserLocal__validar_arquivo_yaml',
+    @mock.patch('vmm_manager.parser.parser_local.ParserLocal._ParserLocal__validar_arquivo_yaml',
                 return_value=None)
     def test_parser_inventario_vms_sem_imagem(self, _, servidor_acesso, monkeypatch):
         dados_teste = DadosTeste()
@@ -43,7 +43,7 @@ class TestParserLocal(Base):
         assert msg == 'Imagem da VM {} não definida.'.format(
             inventario[0][0]['vms'][0]['nome'])
 
-    @mock.patch('vmm_manager.app.inventario.ParserLocal._ParserLocal__validar_arquivo_yaml',
+    @mock.patch('vmm_manager.parser.parser_local.ParserLocal._ParserLocal__validar_arquivo_yaml',
                 return_value=None)
     def test_parser_inventario_nome_vm_duplicado(self, _, servidor_acesso, monkeypatch):
         dados_teste = DadosTeste()
@@ -72,7 +72,7 @@ class TestParserLocal(Base):
         assert msg == 'VM {} referenciada mais de uma vez no inventário.'.format(
             inventario[0][0]['vms'][0]['nome'])
 
-    @mock.patch('vmm_manager.app.inventario.ParserLocal._ParserLocal__validar_arquivo_yaml',
+    @mock.patch('vmm_manager.parser.parser_local.ParserLocal._ParserLocal__validar_arquivo_yaml',
                 return_value=None)
     def test_parser_inventario_min_padrao(self, _, servidor_acesso, monkeypatch):
         dados_teste = DadosTeste()
@@ -100,7 +100,7 @@ class TestParserLocal(Base):
         assert status is True
         assert inventario_resposta == self.get_obj_inventario(inventario)
 
-    @mock.patch('vmm_manager.app.inventario.ParserLocal._ParserLocal__validar_arquivo_yaml',
+    @mock.patch('vmm_manager.parser.parser_local.ParserLocal._ParserLocal__validar_arquivo_yaml',
                 return_value=None)
     def test_parser_inventario_min_sem_padrao(self, _, servidor_acesso, monkeypatch):
         dados_teste = DadosTeste()
